codecs: more docs, a terminology guide, consistency in options.
No breaking changes.  In fact I think this should be behaviorally
identical to before, despite the size of the diff.
There are some functions that are explicitly marked as deprecated now,
but I am not removing them in this diff.

One exception: some struct names did change.  However, these have not
been included in any tagged release so far, so I believe changing them
now is still pretty fair game, and the number of affected people should
be slim to none.

The main source of changes is pushing towards nomenclature consistency,
which is described in the new readme file also included the diff.