selector: add tests which are driven by language-agnostic spec fixtures.
This has been LONG overdue.

It depends on a matching PR to the specs, which is:
https://github.com/ipld/ipld/pull/130 .

It adds the specs+docs repo as a git submodule.
(This may be tetchy to some people, but they Solve Problems.)
If you don't check out the submodule, these tests are just skipped.

It adds github.com/warpfork/go-testmark as a new dependency.
This is what parses the fixture format.  It's nothing special;
the fixture format is also... markdown.  This parser just gives us
a good combination of human readability and machine processability.