Alias a few more things back into the root package.
Many of these will be transitional only, I think,
but we'll see what the appropriate timescale is for that.
It may be fine to leave them for quite a while.

With these aliases, I've been able to substitute this branch in
to... approximately all of our downstreams I know about,
and all their tests still pass.  That includes packages that
have codegen outputs committed already (!).  That includes go-ipfs,
and thus includes all of the transitives between us and it.

I think that means we're truly able to claim this will be a
"no breaking changes" scenario when we merge this and
it next comes time to tag a release.

(Golang's alias feature is kind of incredible.  I'm really happy.)