Update all our packages to no longer reference the root package.
We can do all our work without it, necessarily.  Let's do so.

This could have happened opportunistically, as it was forced, too --
in fact, that's what the prior commits were all doing --
and we'd just expect that to happen more and more as we add new
features and helper methods and glue into the root package.

However, it's just as easy (in fact, probably easier) to do it
proactively, all at once, right now.