Add just a few more aliases.
These are based on manual review of the change in the godoc's list
of exported symbols.
(I haven't actually seen them used anywhere in downstreams in a way
that their omission is a breaking change, but there's always a chance.)

DeepEqual actually completely belongs here.

ADL... I'm dubious on.  I don't really think that should be in the
datamodel package, either, actually.  But let's roll with it.

ErrHashMismatch... well, all those other errors already coming along
for "a transitionary period", so we might as well apply that pattern
consistently, and include this one too.