bindnode: critical that union member setting happens before finisher callbacks!
Amusingly, you only notice the difference when there's a union
(and either a kinded or a stringjoin one at that (so far anyway))
that's a value in a map.

This is because in all other cases, memory sideeffects are sufficient.
But when you're working on a map value -- the finisher callbacks
are actually important, and will cause copying action.
Which means continued writing into the memory you were using up until
that point becomes unimpactful.

(Took me about 2 hours to find these.  Computers.)