bindnode: don't go questing for a representation node just to switch back up to type level.
With this fix, together with the fixes in the prior commit,
encoding a stringprefix union now works correctly.

(And, I guess, a stringprefix union that's directly inside a kinded union,
and that contains a struct with stringjoin representation, all works.
Which, though it sounds like a mouthful, is in fact what I'm doing
in the downstream application that's causing me to fix all this stuff.)