bindnode: fix for stringjoin struct emission when first field is the empty string.
In the case of an empty string as the first field, the buffer length
is not a valid proxy for whether we're on the first field or not.

This means if we have some type like:
`type Foo struct {a String; b String} representation stringjoin(":")`,
and the value of it is `{"", "b"}`, then the string of that
should still be ":b".
Before this fix, it would incorrectly be emitted as "b" (no joiner),
which would not round-trip.

Includes regression test.