codec: make cbor and json codecs use ErrUnexpectedEOF
In particular, if the input doesn't have a single token, we would return
io.EOF, which can be easily mistaken by an error that can be ignored.
In particular, go-ipfs-cmds ignored it as such.

Make them return a clearly bad "unexpected EOF" error,
which mirrors the behavior found in many other unmarshal APIs
such as encoding/json.Unmarshal.

Finally, add a table-driven test that we can extend in the future,
which tests all four codecs.

For ipfs/go-ipfs#8378.