schema/dmt: redesign with bindnode and add Compile
The schema-schema DMT types are now straightforward Go types,
which match the schema definition as per the bindnode rules.

To simplify the process, bindnode now knows how to generate those types.
We're using that via a hacky build-tagged "go test" line,
since otherwise we would have to export the schema to another package.

We delete the schema2 package, as it is now an abandonex experiment,
and its reliance on the old schemadmt types breaks it badly.
Future PRs might resurrect parts of its code inside /schema.

Finally, we add a Compile API to turn a schema DMT node
into a set of types registered in a TypeSystem.

While at it, we also update the schema-schema in schemadmt
to be a bit closer to the spec with the type names.

Updates #188.