schema: add ways to set and access the ImplicitValue for a struct field.
And make that 'union' type actually work in the golang side.

And add bools to it, which were missing.

And stop using struct wrappers; I don't think there's any reason
to care about someone casting things into those types.

This may not be an ideal API for this.  It should work for today,
however, and that's good enough.  (We may be able to start revisiting
more design choices in this area after the nest generation of
schema dmt and parsing work lands (which will be *very* soon now);
but we'll see.)