schema: turn TypeName into an alias
The named type doesn't gain us anything.
To make it truly safe from misuse, we'd need to make it opaque.
That seems overkill though, and it would add red tape.

Instead, we should continue to rely on typesystem validation
and the upcoming parser to verify that type names are valid.

At that point, the only remaining use for TypeName is godocs,
to clarify when string parameters are type names.
For that purpose, a type alias is just fine.

The big advantage of moving to a type alias is the easier code.

Technically a breaking change, if any downstream was calling String.
However, most strings shouldn't be fiddling with the schema package,
and the fix is trivially easy: just remove ".String()". For example:

    sed -i 's/\.Name()\.String()/.Name()/' **/*.go