Round-trip test nodes through custom codec to ensure stability
dag-json has lexical sorting of map fields, so encode() here will be reordering
some of the fields. Since we are doing a traversal from rootNode for most tests
and it doesn't get loaded (since it's the start Node), it doesn't get the same
reordering treatment as the other blocks.

So, we round-trip all blocks through the codec to ensure that the Node that
we have matches the Link and is what it would be regardless of where it comes
from.

Additionally, we're using a custom codec here for the tests that _remove_ the
map sorting and leaves the blocks unmolested, so we don't have to rewrite all
of the parts of the tests where ordering matters and the fields match up
more nicely with the fixtures we've constructed.

For the chain test, the ordering is now different because it depended on
dag-json reordering, so instead of hard-wiring expectations in a switch() we
do it in an array of expected paths.