cli: add new packages (and new go.mod) for a CLI.
These are still outlines but begin to show the direction I want to go.

The use of a new go module means the rest of the go-ipld-prime library
can still be used without drawing in any new transitive dependencies
on CLI libraries, etc.

The CLI library used is the urfave/cli package.  It works.
(I don't know if I'm thrilled with it -- especially with what it
provides for usage and help text formatting defaults -- but it works.)

A commit from the very tip of go-testmark is used, because
it introduces new features for testing scripts and executables.
These are used to test the commands of this new baby CLI app!
Right now, it only tests that the help text for one of the
subcommands is what I think it is... but there's room to grow.
(Also, that's turned out surprisingly useful, because it lets me see
a quick diff of what changes in the help text when I'm editing!)

Using an argument of "-testmark.regen" to "go test" will also cause
the fixtures to be *automatically updated*.  Isn't that nice?

Due to a lack of creativity, the working name for the command is
simply "ipld".  This may change.  (To compare and contrast: consider
that the 'jq' tool is not called 'json'.  So a CLI tool for working
with IPLD similiarly doesn't necessarily need to be named 'ipld'.)
For now, a working title is needed, so here we go.