cli: implement a first path through the read command.
There are many usages which still result in panics for unimplemented features,
but at least one demo works now: when you pipe through your own data.

More importantly: the testing system is keeping up.
We now have a test which demonstrates how to use piping behavior!
(This relies on bash existing on the host and generally acting
"roughly like one would expect", which means it's probably not working
on windows.  That said... it's probably still a pretty useful trade.)