drop dagjson2.
This work never got enough attention block to be carried to completion.
As such, it's not carrying enough weight to be worth keeping in tree.

If we do a refactor and replacement of the json guts in the future,
it'll almost certainly be by using the encoder and decoder features
of https://pkg.go.dev/github.com/go-json-experiment/json , which
appears to be nicely decoupled and reusable and should be
able to do all the right things.