drop codectools.
Same as with the dagjson2: this didn't get enough attention allocated
to it to go places.  We might still want to revisit some of the guts
of implementation of the bundled json and cbor codecs someday, but
if so, we'll restart at that time.  Right now, this code is not
load-bearing, so keeping it in-tree is unjustifiable.