cli: 'schema parse' command largely works.
go-testmark library version bumped to get the new filesystem helpers,
which help us write this fixture cleanly.

Some of my initally intended DSL content is commented back out for now,
because it causes a panic.  I haven't tracked it down yet, but since
it's going to be either a parser bug or a bindnode bug, the fix for it
should probably be handled on some other more focused PR.