cli: Drop entire dir.  Extracting to a new repo.
Trying to use multiple go modules in a repo was an experiment.
I haven't been thrilled with the results.
(It works?  But it's a stretch to say it works _well_.
The UX can easily become confusing; it still requires multiple
commits to make changes in the parent and the downstream module,
which is just as bad as separate repos; and rebasing the parent
becomes impossible after the downstream refers to it; in general,
it's as much logistical overhead as a separate repo, but without
the *clarity* of a separate repo, so it's just not a win.)

All this code is still dandy; I'm going to move it to a new repo
at https://github.com/ipld/go-ipldtool/ .