add binary access to the interface for Link.
This is a pretty big change if anyone else out there
is implementing new link types.

If you're using the linking/cid package, the change is already done.

I'm going to consider this a minimally breaking change because
I don't think we've got a lot of diverse Link implementations
out in the wild.

(And even if so: this should be a pretty easy addition to make.)

I'm anticipating using this as part of a good layering of APIs
regarding storage.  (We should be able to use binary, dense things
in any storage substrate that allows it.  Some common storage systems,
like e.g. flatfs, don't work with binary.  But any escaping having to
do with that should be associated with the storage system -- not
pushed to any other layer.  Having binary access here helps do that.)